perf(gazelle): make modules_mapping incremental
Generate a module mapping separately for each wheel and merge the
results to speed up updates. Previously the `module_mapping` rule had to
regenerate the module mapping for all wheels whenever a wheel was
updated. By splitting the generation and merging steps we only need to
rerun the generation for updated wheels and merge the results.