[wip]feat: add module mode to gazelle extension
Add module based generation to the gazelle extension. This allows users
to generate targets at the file, or module, level based on file name and
the presence of `__main__` dunders within a file.

This is a work in progress. Ideally, this should not modify the default
behavior and will be an opt-in change. However, this is not currently true.
See the breaking change below.

BREAKING_CHANGE: The `pyKinds` now match based on sources files using
`MatchAttrs: []string{"srcs"}` instead of `matchAll: true`. This seems
to be required for module-based generation, but breaks one of the examples.